airflow/requirements-airflow.txt
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
events==0.5
executing==2.2.1
grpcio==1.78.0
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jupyter-client==8.8.0
jupyter-core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
opensearch-protobufs==0.19.0
opensearch-py==3.1.0
packaging==26.0
parso==0.8.6
platformdirs==4.9.4
prompt-toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
psycopg2-binary==2.9.11
pure-eval==0.2.3
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyzmq==27.1.0
requests==2.32.5
six==1.17.0
stack-data==0.6.3
tornado==6.5.5
traitlets==5.14.3
typing-extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
